fix: renamed `RegisterBanner` component to `AuthBanner`
