+++ resources/js/components/Home/AuthBanner.vue
